Prefer VISUAL over EDITOR to specify editor
With this commit, pick uses the programme specified by the VISUAL
environment variable to edit notes. If VISUAL is not set, it will fall
back to the previous behaviour of using the programme specified by the
EDITOR environment variable, or vim if EDITOR is not set. This behaviour
is implemented in a new editorPath function extracted from
editorReadText.